refactor(window): move send_lsp_changes_for_buffer to impl Window
The body only touches per-window state — buffer_metadata, the
LSP manager, debounce schedules, and the buffers map all live on
Window. Removing the active_window indirection lets the function
operate directly on its owning fields and gives callers a path
to skip the Editor multiplexer.

Editor-side wrappers were just `self.active_window_mut().X(...)`
forwards; switched call sites in event_apply, lsp_requests,
plugin_commands, recovery_actions to call the Window method
directly. Drops the now-unused INLAY_HINTS_DEBOUNCE_MS const
that lived on lsp_requests's Editor impl block.
