+++ crates/fresh-editor/src/app/event_apply.rs
@@ -175,10 +175,14 @@ impl Editor {
                     range_length: None,
                     text: full_text,
                 }];
-                self.send_lsp_changes_for_buffer(self.active_buffer(), full_change);
+                let buf = self.active_buffer();
+                self.active_window_mut()
+                    .send_lsp_changes_for_buffer(buf, full_change);
             }
         } else {
-            self.send_lsp_changes_for_buffer(self.active_buffer(), lsp_changes);
+            let buf = self.active_buffer();
+            self.active_window_mut()
+                .send_lsp_changes_for_buffer(buf, lsp_changes);
         }
     }
 
@@ -510,7 +514,8 @@ impl Editor {
             })
             .unwrap_or_default();
         if !full_content_change.is_empty() {
-            self.send_lsp_changes_for_buffer(buffer_id, full_content_change);
+            self.active_window_mut()
+                .send_lsp_changes_for_buffer(buffer_id, full_content_change);
         }
 
         Some(bulk_edit)
