+++ crates/fresh-editor/src/app/lsp_requests.rs
@@ -62,11 +62,6 @@ fn lsp_range_contains(range: &lsp_types::Range, line: u32, character: u32) -> bo
 
 const SEMANTIC_TOKENS_RANGE_DEBOUNCE_MS: u64 = 50;
 const SEMANTIC_TOKENS_RANGE_PADDING_LINES: usize = 10;
-/// Debounce window between the last buffer edit and the next inlay hints
-/// re-request. Matches the diagnostic-pull debounce to keep network chatter
-/// low while still refreshing hints after brief editing pauses (including
-/// saves, which naturally follow an edit).
-const INLAY_HINTS_DEBOUNCE_MS: u64 = 500;
 
 impl Editor {
     /// Handle LSP completion response.
@@ -2905,222 +2900,10 @@ impl Editor {
         }
 
         // Notify LSP about the changes using pre-calculated positions
-        self.send_lsp_changes_for_buffer(buffer_id, lsp_changes);
+        self.active_window_mut()
+            .send_lsp_changes_for_buffer(buffer_id, lsp_changes);
 
         Ok(())
-    }
-
-    /// Send pre-calculated LSP changes for a specific buffer
-    pub(crate) fn send_lsp_changes_for_buffer(
-        &mut self,
-        buffer_id: BufferId,
-        changes: Vec<TextDocumentContentChangeEvent>,
-    ) {
-        if changes.is_empty() {
-            return;
-        }
-
-        // Check if LSP is enabled for this buffer
-        let metadata = match self.active_window().buffer_metadata.get(&buffer_id) {
-            Some(m) => m,
-            None => {
-                tracing::debug!(
-                    "send_lsp_changes_for_buffer: no metadata for buffer {:?}",
-                    buffer_id
-                );
-                return;
-            }
-        };
-
-        if !metadata.lsp_enabled {
-            tracing::debug!("send_lsp_changes_for_buffer: LSP disabled for this buffer");
-            return;
-        }
-
-        // Get the URI
-        let uri = match metadata.file_uri() {
-            Some(u) => u.clone(),
-            None => {
-                tracing::debug!(
-                    "send_lsp_changes_for_buffer: no URI for buffer (not a file or URI creation failed)"
-                );
-                return;
-            }
-        };
-        let file_path = metadata.file_path().cloned();
-
-        // Get language from buffer state
-        let language = match self
-            .windows
-            .get(&self.active_window)
-            .map(|w| &w.buffers)
-            .expect("active window present")
-            .get(&buffer_id)
-            .map(|s| s.language.clone())
-        {
-            Some(l) => l,
-            None => {
-                tracing::debug!(
-                    "send_lsp_changes_for_buffer: no buffer state for {:?}",
-                    buffer_id
-                );
-                return;
-            }
-        };
-
-        tracing::trace!(
-            "send_lsp_changes_for_buffer: sending {} changes to {} in single didChange notification",
-            changes.len(),
-            uri.as_str()
-        );
-
-        // Check if we can use LSP (respects auto_start setting)
-        use crate::services::lsp::manager::LspSpawnResult;
-        let __active_id = self.active_window;
-        let Some(__win) = self.windows.get_mut(&__active_id) else {
-            tracing::debug!("send_lsp_changes_for_buffer: no active window");
-            return;
-        };
-        let Some(lsp) = __win.lsp.as_mut() else {
-            tracing::debug!("send_lsp_changes_for_buffer: no LSP manager available");
-            return;
-        };
-
-        if lsp.try_spawn(&language, file_path.as_deref()) != LspSpawnResult::Spawned {
-            tracing::debug!(
-                "send_lsp_changes_for_buffer: LSP not running for {} (auto_start disabled)",
-                language
-            );
-            return;
-        }
-
-        // Check which handles need didOpen first
-        let handles_needing_open: Vec<_> = {
-            let Some(metadata) = __win.buffer_metadata.get(&buffer_id) else {
-                return;
-            };
-            lsp.get_handles(&language)
-                .into_iter()
-                .filter(|sh| !metadata.lsp_opened_with.contains(&sh.handle.id()))
-                .map(|sh| (sh.name.clone(), sh.handle.id()))
-                .collect()
-        };
-
-        if !handles_needing_open.is_empty() {
-            // Get text for didOpen
-            let text = match self
-                .buffers()
-                .get(&buffer_id)
-                .and_then(|s| s.buffer.to_string())
-            {
-                Some(t) => t,
-                None => {
-                    tracing::debug!(
-                        "send_lsp_changes_for_buffer: buffer text not available for didOpen"
-                    );
-                    return;
-                }
-            };
-
-            // Send didOpen to all handles that haven't been opened yet
-            let __active_id = self.active_window;
-            let Some(__win) = self.windows.get_mut(&__active_id) else {
-                return;
-            };
-            let Some(lsp) = __win.lsp.as_mut() else {
-                return;
-            };
-            for sh in lsp.get_handles_mut(&language) {
-                if handles_needing_open
-                    .iter()
-                    .any(|(_, id)| *id == sh.handle.id())
-                {
-                    if let Err(e) =
-                        sh.handle
-                            .did_open(uri.as_uri().clone(), text.clone(), language.clone())
-                    {
-                        tracing::warn!(
-                            "Failed to send didOpen to '{}' before didChange: {}",
-                            sh.name,
-                            e
-                        );
-                    } else {
-                        tracing::debug!(
-                            "Sent didOpen for {} to LSP handle '{}' before didChange",
-                            uri.as_str(),
-                            sh.name
-                        );
-                    }
-                }
-            }
-
-            // Mark all as opened
-            if let Some(metadata) = __win.buffer_metadata.get_mut(&buffer_id) {
-                for (_, handle_id) in &handles_needing_open {
-                    metadata.lsp_opened_with.insert(*handle_id);
-                }
-            }
-
-            // didOpen already contains the full current buffer content, so we must
-            // NOT also send didChange (which carries pre-edit incremental changes).
-            // Sending both would corrupt the server's view of the document.
-            return;
-        }
-
-        // Now send didChange to all handles for this language
-        let __active_id = self.active_window;
-        let Some(lsp) = self
-            .windows
-            .get_mut(&__active_id)
-            .and_then(|w| w.lsp.as_mut())
-        else {
-            return;
-        };
-        let mut any_sent = false;
-        for sh in lsp.get_handles_mut(&language) {
-            if let Err(e) = sh.handle.did_change(uri.as_uri().clone(), changes.clone()) {
-                tracing::warn!("Failed to send didChange to '{}': {}", sh.name, e);
-            } else {
-                any_sent = true;
-            }
-        }
-        if any_sent {
-            tracing::trace!("Successfully sent batched didChange to LSP");
-
-            // Invalidate diagnostic cache so the next diagnostic apply recomputes
-            // overlay positions from fresh byte offsets (the buffer content changed)
-            if let Some(state) = self
-                .windows
-                .get(&self.active_window)
-                .map(|w| &w.buffers)
-                .expect("active window present")
-                .get(&buffer_id)
-            {
-                if let Some(path) = state.buffer.file_path() {
-                    crate::services::lsp::diagnostics::invalidate_cache_for_file(
-                        &path.to_string_lossy(),
-                    );
-                }
-            }
-
-            // Schedule debounced diagnostic re-pull (1000ms after last edit)
-            self.active_window_mut().scheduled_diagnostic_pull = Some((
-                buffer_id,
-                std::time::Instant::now() + std::time::Duration::from_millis(1000),
-            ));
-
-            // Schedule debounced inlay hints refresh. Without this, hints
-            // computed before the edit remain anchored to stale byte offsets
-            // (including inside ranges the user just deleted), and new hints
-            // that the server would now produce never arrive.
-            if self.config.editor.enable_inlay_hints {
-                self.active_window_mut().scheduled_inlay_hints_request = Some((
-                    buffer_id,
-                    std::time::Instant::now()
-                        + std::time::Duration::from_millis(INLAY_HINTS_DEBOUNCE_MS),
-                ));
-            }
-        }
     }
 
     /// Start rename mode - select the symbol at cursor and allow inline editing
